Some slight refactoring.
